if the plug-in is in a different process group than GIMP, kill the entire
2006-09-20  Michael Natterer  <mitch@gimp.org>

	* app/plug-in/gimpplugin.c (gimp_plug_in_close): if the plug-in is
	in a different process group than GIMP, kill the entire group so
	the plug-in's children are killed too.

	* plug-ins/uri/uri-backend-wget.c (uri_backend_load_image): open
	a process group so the wget process gets killed (bug #356643).

+++ app/plug-in/gimpplugin.c
@@ -20,6 +20,8 @@
 
 #include "config.h"
 
+#define _GNU_SOURCE
+
 #include <errno.h>
 #include <signal.h>
 #include <stdlib.h>
@@ -427,7 +429,13 @@ gimp_plug_in_close (GimpPlugIn *plug_in,
             g_print (_("Terminating plug-in: '%s'\n"),
                      gimp_filename_to_utf8 (plug_in->prog));
 
-          status = kill (plug_in->pid, SIGKILL);
+          /*  If the plug-in opened a process group, kill the group instead
+           *  of only the plug-in, so we kill the plug-in's children too
+           */
+          if (getpgid (0) != getpgid (plug_in->pid))
+            status = kill (- plug_in->pid, SIGKILL);
+          else
+            status = kill (plug_in->pid, SIGKILL);
         }
 
       /* Wait for the process to exit. This will happen
